sanitizePath plugin unit tests
* Remove unused utils.santizePath implementation
* Remove test for utils.sanitizePath
* Add test for sanitizePath plugin
