+++ lib/utils.js
@@ -2,31 +2,6 @@
 
 'use strict';
 
-var assert = require('assert-plus');
-
-
-/**
- * Cleans up sloppy URL paths, like /foo////bar/// to /foo/bar.
- * @public
- * @method  sanitizePath
- * @param   {String} path the HTTP resource path.
- * @returns {String}      Cleaned up form of path.
- */
-function sanitizePath(path) {
-    assert.ok(path);
-
-    // Be nice like apache and strip out any //my//foo//bar///blah
-    path = path.replace(/\/\/+/g, '/');
-
-    // Kill a trailing '/'
-    if (path.lastIndexOf('/') === (path.length - 1) && path.length > 1) {
-        path = path.substr(0, path.length - 1);
-    }
-
-    return (path);
-}
-
-
 /**
  * Return a shallow copy of the given object;
  * @public
@@ -83,7 +58,6 @@ function mergeQs(obj1, obj2) {
 ///--- Exports
 
 module.exports = {
-    sanitizePath: sanitizePath,
     shallowCopy: shallowCopy,
     mergeQs: mergeQs
 };
