+++ test/utils.test.js
@@ -2,7 +2,6 @@
 
 var utils = require('../lib/utils.js');
 var mergeQs = utils.mergeQs;
-var sanitizePath = utils.sanitizePath;
 
 if (require.cache[__dirname + '/lib/helper.js']) {
     delete require.cache[__dirname + '/lib/helper.js'];
@@ -29,33 +28,6 @@ test('merge qs', function (t) {
 
     var qs5 = mergeQs(null, null);
     t.deepEqual(qs5, {});
-
-    t.done();
-});
-
-test('sanitizePath', function (t) {
-    // Ensure it santizies potential edge cases correctly
-    var tests = {
-        input: [
-            '////foo////', //excess padding on both ends
-            'bar/foo/', // trailing slash
-            'bar/foo/////', // multiple trailing slashes
-            'foo////bar', // multiple slashes inbetween
-            '////foo', // multiple at beginning
-            '/foo/bar' // don't mutate
-        ],
-        output: [
-            '/foo',
-            'bar/foo',
-            'foo/bar',
-            '/foo',
-            '/foo/bar'
-        ]
-    };
-
-    for (var i = 0; i < tests; i++) {
-        t.equal(sanitizePath(tests.input[i]), tests.output[i]);
-    }
 
     t.done();
 });
